Crate ark_ff

Re-exports

pub use self::bytes::*;

pub use self::biginteger::*;

pub use self::fields::*;

Modules

biginteger

bytes

fields

prelude

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

field_new

impl_additive_ops_from_ref

impl_multiplicative_ops_from_ref

push_to_vec

to_bytes

Takes as input a sequence of structs, and converts them to a series of bytes. All traits that implement Bytes can be automatically converted to bytes in this manner.

vec

Creates a Vec containing the arguments.

Traits

One

Defines a multiplicative identity element for Self.

ToConstraintField

Types that can be converted to a vector of F elements. Useful for specifying how public inputs to a constraint system should be represented inside that constraint system.

UniformRand

Zero

Defines an additive identity element for Self.