Module ark_ff::fields

Re-exports

pub use ark_ff_macros;

pub use self::models::*;

Modules

arithmetic

macros

models

utils

Structs

BitIteratorBE

Iterates over a slice of u64 in big-endian order.

BitIteratorLE

Iterates over a slice of u64 in little-endian order.

Enums

LegendreSymbol

Traits

FftField

The interface for fields that are able to be used in FFTs.

FftParameters

A trait that defines parameters for a field that can be used for FFTs.

Field

The interface for a generic field.

FpParameters

A trait that defines parameters for a prime field.

PrimeField

The interface for a prime field.

SquareRootField

The interface for a field that supports an efficient square-root operation.

Functions

batch_inversion

batch_inversion_and_mul