ark_ff::fields
Module models

Re-exports

pub use self::fp2::*;

pub use self::fp3::*;

pub use self::fp4::*;

pub use self::fp6_3over2::*;

pub use self::fp12_2over3over2::*;

pub use quadratic_extension::*;

pub use cubic_extension::*;

Modules

cubic_extension

fp2

fp3

fp4

fp6_2over3

fp6_3over2

fp12_2over3over2

quadratic_extension

Structs

Fp64

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 64 bits.

Fp256

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 256 bits.

Fp320

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 320 bits.

Fp384

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 384 bits.

Fp448

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 448 bits.

Fp768

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 768 bits.

Fp832

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most 832 bits.

Traits

Fp64Parameters

Fp256Parameters

Fp320Parameters

Fp384Parameters

Fp448Parameters

Fp768Parameters

Fp832Parameters