ark_ff::fields::models::cubic_extension

Struct CubicExtField

pub struct CubicExtField<P: CubicExtConfig> {
    pub c0: P::BaseField,
    pub c1: P::BaseField,
    pub c2: P::BaseField,
}

An element of a cubic extension field F_p[X]/(X^3 - P::NONRESIDUE) is represented as c0 + c1 * X + c2 * X^2, for c0, c1, c2 in P::BaseField.

Fields
c0: P::BaseFieldc1: P::BaseFieldc2: P::BaseField
Implementations

impl<P: Fp3Config> CubicExtField<Fp3ConfigWrapper<P>>

pub fn mul_assign_by_fp(&mut self, value: &P::Fp)

In-place multiply all coefficients c0, c1, and c2 of self by an element from Fp.

Examples

use ark_mnt6_753::{Fq as Fp, Fq3 as Fp3};
let c0: Fp = Fp::rand(&mut test_rng());
let c1: Fp = Fp::rand(&mut test_rng());
let c2: Fp = Fp::rand(&mut test_rng());
let mut ext_element: Fp3 = Fp3::new(c0, c1, c2);

let base_field_element: Fp = Fp::rand(&mut test_rng());
ext_element.mul_assign_by_fp(&base_field_element);

assert_eq!(ext_element.c0, c0 * base_field_element);
assert_eq!(ext_element.c1, c1 * base_field_element);
assert_eq!(ext_element.c2, c2 * base_field_element);

impl<P: Fp6Config> CubicExtField<Fp6ConfigWrapper<P>>

pub fn mul_assign_by_fp2(&mut self, other: Fp2<P::Fp2Config>)

pub fn mul_by_fp(&mut self, element: &<P::Fp2Config as Fp2Config>::Fp)

pub fn mul_by_fp2(&mut self, element: &Fp2<P::Fp2Config>)

pub fn mul_by_1(&mut self, c1: &Fp2<P::Fp2Config>)

pub fn mul_by_01(&mut self, c0: &Fp2<P::Fp2Config>, c1: &Fp2<P::Fp2Config>)

impl<P: CubicExtConfig> CubicExtField<P>

pub const fn new(c0: P::BaseField, c1: P::BaseField, c2: P::BaseField) -> Self

Create a new field element from coefficients c0, c1 and c2 so that the result is of the form c0 + c1 * X + c2 * X^2.

Examples

use ark_ff::models::cubic_extension::CubicExtField;

let c0: Fp2 = Fp2::rand(&mut test_rng());
let c1: Fp2 = Fp2::rand(&mut test_rng());
let c2: Fp2 = Fp2::rand(&mut test_rng());
// `Fp6` a degree-3 extension over `Fp2`.
let c: CubicExtField<Config> = Fp6::new(c0, c1, c2);

pub fn mul_assign_by_base_field(&mut self, value: &P::BaseField)

pub fn norm(&self) -> P::BaseField

Calculate the norm of an element with respect to the base field P::BaseField. The norm maps an element a in the extension field Fq^m to an element in the BaseField Fq. Norm(a) = a * a^q * a^(q^2)