ark_ff::fields::models::cubic_extension
Trait CubicExtParameters

pub trait CubicExtParameters:
    'static
    + Send
    + Sync {
    type BasePrimeField: PrimeField;
    type BaseField: Field<BasePrimeField = Self::BasePrimeField>;
    type FrobCoeff: Field;

    const DEGREE_OVER_BASE_PRIME_FIELD: usize;
    const NONRESIDUE: Self::BaseField;
    const FROBENIUS_COEFF_C1: &'static [Self::FrobCoeff];
    const FROBENIUS_COEFF_C2: &'static [Self::FrobCoeff];

    // Required method
    fn mul_base_field_by_frob_coeff(
        c1: &mut Self::BaseField,
        c2: &mut Self::BaseField,
        power: usize,
    );

    // Provided method
    fn mul_base_field_by_nonresidue(fe: &Self::BaseField) -> Self::BaseField { ... }
}

Required Associated Constants

const DEGREE_OVER_BASE_PRIME_FIELD: usize

The degree of the extension over the base prime field.

const NONRESIDUE: Self::BaseField

The cubic non-residue used to construct the extension.

const FROBENIUS_COEFF_C1: &'static [Self::FrobCoeff]

Coefficients for the Frobenius automorphism.

const FROBENIUS_COEFF_C2: &'static [Self::FrobCoeff]

Required Associated Types

type BasePrimeField: PrimeField

The prime field that this cubic extension is eventually an extension of.

type BaseField: Field<BasePrimeField = Self::BasePrimeField>

The base field that this field is a cubic extension of.

type FrobCoeff: Field

The type of the coefficients for an efficient implemntation of the Frobenius endomorphism.

Required Methods

fn mul_base_field_by_frob_coeff( c1: &mut Self::BaseField, c2: &mut Self::BaseField, power: usize, )

A specializable method for multiplying an element of the base field by the appropriate Frobenius coefficient.

Provided Methods

fn mul_base_field_by_nonresidue(fe: &Self::BaseField) -> Self::BaseField

A specializable method for multiplying an element of the base field by the quadratic non-residue. This is used in multiplication and squaring.

Object Safety

This trait is not object safe.