ark_ff::fields::models::fp2

Type Alias Fp2

pub type Fp2<P> = QuadExtField<Fp2ConfigWrapper<P>>;

Alias for instances of quadratic extension fields. Helpful for omitting verbose instantiations involving Fp2ConfigWrapper.

Aliased Type

struct Fp2<P> {
    pub c0: <Fp2ConfigWrapper<P> as QuadExtConfig>::BaseField,
    pub c1: <Fp2ConfigWrapper<P> as QuadExtConfig>::BaseField,
}

Fields
c0: <Fp2ConfigWrapper<P> as QuadExtConfig>::BaseField

Coefficient c0 in the representation of the field element c = c0 + c1 * X

c1: <Fp2ConfigWrapper<P> as QuadExtConfig>::BaseField

Coefficient c1 in the representation of the field element c = c0 + c1 * X

Implementations

impl<P: Fp2Config> Fp2<P>

pub fn mul_assign_by_fp(&mut self, other: &P::Fp)

In-place multiply both coefficients c0 and c1 of self by an element from Fp.

Examples

let c0: Fp = Fp::rand(&mut test_rng());
let c1: Fp = Fp::rand(&mut test_rng());
let mut ext_element: Fp2 = Fp2::new(c0, c1);

let base_field_element: Fp = Fp::rand(&mut test_rng());
ext_element.mul_assign_by_fp(&base_field_element);

assert_eq!(ext_element.c0, c0 * base_field_element);
assert_eq!(ext_element.c1, c1 * base_field_element);