ark_ff::fields::models::fp6_2over3

Type Alias Fp6

pub type Fp6<P> = QuadExtField<Fp6ConfigWrapper<P>>;

Aliased Type

struct Fp6<P> {
    pub c0: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField,
    pub c1: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField,
}

Fields
c0: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField

Coefficient c0 in the representation of the field element c = c0 + c1 * X

c1: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField

Coefficient c1 in the representation of the field element c = c0 + c1 * X

Implementations

impl<P: Fp6Config> Fp6<P>

pub fn mul_by_034( &mut self, c0: &<P::Fp3Config as Fp3Config>::Fp, c3: &<P::Fp3Config as Fp3Config>::Fp, c4: &<P::Fp3Config as Fp3Config>::Fp, )

pub fn mul_by_014( &mut self, c0: &<P::Fp3Config as Fp3Config>::Fp, c1: &<P::Fp3Config as Fp3Config>::Fp, c4: &<P::Fp3Config as Fp3Config>::Fp, )