ark_ff
Macro BigInt

macro_rules! BigInt {
    ($c0:expr) => { ... };
}

Construct a BigInt<N> element from a literal string.

Panics

If the integer represented by the string cannot fit in the number of limbs of the BigInt, this macro results in a

compile-time error if used in a const context

run-time error otherwise.

Usage

const ONE: BigInt<6> = BigInt!("1");

fn check_correctness() {
    assert_eq!(ONE, BigInt::from(1u8));
}