ark_ff::fields::models::fp

Struct Fp

pub struct Fp<P: FpConfig<N>, const N: usize>(/* private fields */);

Represents an element of the prime field F_p, where p == P::MODULUS. This type can represent elements in any field of size at most N * 64 bits.

Implementations

impl<T: MontConfig<N>, const N: usize> Fp<MontBackend<T, N>, N>

pub const fn new(element: BigInt<N>) -> Self

Construct a new field element from its underlying BigInt data type.

pub const fn new_unchecked(element: BigInt<N>) -> Self

Construct a new field element from its underlying BigInt data type.

Unlike Self::new, this method does not perform Montgomery reduction. Thus, this method should be used only when constructing an element from an integer that has already been put in Montgomery form.