Trait ark_ff::ToConstraintField

pub trait ToConstraintField<F: Field> {
    fn to_field_elements(&self) -> Option<Vec<F>>;
}

Types that can be converted to a vector of F elements. Useful for specifying how public inputs to a constraint system should be represented inside that constraint system.

Required Methods

fn to_field_elements(&self) -> Option<Vec<F>>